+++ redbot/cogs/audio/cog/utils.py
@@ -1,13 +1,15 @@
 from abc import ABC
 from pathlib import Path
 
+from redbot import VersionInfo, version_info
 from redbot.core import commands
 from redbot.core.i18n import Translator
-
 from ..converters import get_lazy_converter, get_playlist_converter
 
+red_extras = version_info.to_json()
+red_extras.update({"major": 2, "minor": 0, "micro": 0})
+__version__ = VersionInfo.from_json(red_extras)
 
-__version__ = "1.1.0"
 __author__ = ["aikaterna", "Draper"]
 
 _ = Translator("Audio", Path(__file__).parent)
